+++ src/vm-memory/src/mmap.rs
@@ -784,14 +784,14 @@ mod tests {
 
     #[test]
     fn test_bitmap_update_on_write() {
-        let page_size = 4096 as usize;
+        let page_size = 4096_usize;
         let mut mmap =
             GuestRegionMmap::new(MmapRegion::new(page_size * 5).unwrap(), GuestAddress(0x0))
                 .unwrap();
         mmap.enable_dirty_page_tracking();
 
         // check write_obj
-        let sample_val = 0xaa55_aa55_aa55_aa55 as u64;
+        let sample_val = 0xaa55_aa55_aa55_aa55_u64;
         assert!(!mmap.dirty_bitmap().unwrap().is_addr_set(0));
         assert!(mmap.write_obj(sample_val, MemoryRegionAddress(0)).is_ok());
         assert!(mmap.dirty_bitmap().unwrap().is_addr_set(0));
@@ -971,10 +971,7 @@ mod tests {
 
     #[test]
     fn test_overlapping_memory_regions() {
-        let regions_summary = [
-            (GuestAddress(0), 100 as usize),
-            (GuestAddress(99), 100 as usize),
-        ];
+        let regions_summary = [(GuestAddress(0), 100_usize), (GuestAddress(99), 100_usize)];
 
         assert_eq!(
             format!(
@@ -1027,10 +1024,7 @@ mod tests {
 
     #[test]
     fn test_unsorted_memory_regions() {
-        let regions_summary = [
-            (GuestAddress(100), 100 as usize),
-            (GuestAddress(0), 100 as usize),
-        ];
+        let regions_summary = [(GuestAddress(100), 100_usize), (GuestAddress(0), 100_usize)];
 
         assert_eq!(
             format!(
@@ -1083,10 +1077,7 @@ mod tests {
 
     #[test]
     fn test_valid_memory_regions() {
-        let regions_summary = [
-            (GuestAddress(0), 100 as usize),
-            (GuestAddress(100), 100 as usize),
-        ];
+        let regions_summary = [(GuestAddress(0), 100_usize), (GuestAddress(100), 100_usize)];
 
         let guest_mem = GuestMemoryMmap::new();
         assert_eq!(guest_mem.regions.len(), 0);